course data fixes

+++ fs-2025-a-api-demo-002/Models/Course.cs
@@ -9,7 +9,7 @@
         public string courseType { get; set; }
         public string courseName { get; set; }
         public int courseLessonCount { get; set; }
-        public int courseLengthInHours { get; set; }
+        public decimal courseLengthInHours { get; set; }
         public string shortDescription { get; set; }
         public string courseImage { get; set; }
         public int priceInUSD { get; set; }
